+++ Modules/Utils/Sources/PocketCastsUtils/Feature Flags/FeatureFlag.swift
@@ -375,9 +375,9 @@ public enum FeatureFlag: String, CaseIterable {
         case .episodeDetailTranscript:
             true
         case .bannerAdPodcasts:
-            false
+            true
         case .bannerAdPlayer:
-            false
+            true
         case .streamingCustomSessionConfiguration:
             true
         case .guestListsNetworkHighlightsRedesign: