+++ podcasts/Analytics/AnalyticsEvent.swift
@@ -758,6 +758,8 @@ enum AnalyticsEvent: String {
     case recommendationsDismissed
     case recommendationsSearchTapped
     case recommendationsMoreTapped
+    case recommendationsContinueTapped
+    case recommendationsImportTapped
 
     // MARK: - Interests
     case onboardingInterestsShown